+++ Sources/HSLuvSwift/Info.plist
