+++ Sources/HSLuvSwift/Protocols.swift
