+++ Tests/HSLuvSwiftTests/MacExtensions/AppKit.swift
@@ -22,17 +22,34 @@
 // SOFTWARE.
 //
 
-import UIKit
+import Foundation
+import XCTest
+import HSLuvSwift
 
-extension UIColor: HSLuvInitializable, HSLuvConvertible {
+#if canImport(AppKit)
 
-    /// Convenience function to wrap the behavior of getRed(red:green:blue:alpha:)
-    public func getRGB() -> (red: CGFloat, green: CGFloat, blue: CGFloat) {
-        var red: CGFloat = 0, green: CGFloat = 0, blue: CGFloat = 0, alpha: CGFloat = 0
-        getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-        return (red, green, blue)
+class AppKitTests: XCTestCase {
+    let rgbRangeTolerance = 0.00000000001
+
+    func testNSColorRGBRangeTolerance() {
+        for h in stride(from: 0.0, through: 360, by: 5) {
+            for s in stride(from: 0.0, through: 100, by: 5) {
+                for l in stride(from: 0.0, through: 100, by: 5) {
+                    let color = NSColor(hue: h, saturation: s, lightness: l, alpha: 1.0)
+
+                    XCTAssertNotNil(color)
+
+                    let tRgb = color.getRGB()
+                    let rgb = [tRgb.red, tRgb.green, tRgb.blue].map { Double($0) }
+
+                    for channel in rgb {
+                        XCTAssertGreaterThan(channel, -rgbRangeTolerance, "HSLuv: \([h, s, l]) -> RGB: \(rgb)")
+                        XCTAssertLessThanOrEqual(channel, 1 + rgbRangeTolerance, "HSLuv: \([h, s, l]) -> RGB: \(rgb)")
+                    }
+                }
+            }
+        }
     }
-
 }
 
-extension UIColor: HPLuvInitializable, HPLuvConvertible {}
+#endif
