+++ Tests/HSLuvSwiftTests/iOSExtensions/UIKit.swift
@@ -26,14 +26,16 @@ import Foundation
 import XCTest
 import HSLuvSwift
 
-class AppKitTests: XCTestCase {
+#if canImport(UIKit)
+
+class UIKitTests: XCTestCase {
     let rgbRangeTolerance = 0.00000000001
 
-    func testNSColorRGBRangeTolerance() {
-        for h in stride(from: 0.0, through: 360, by: 5) {
-            for s in stride(from: 0.0, through: 100, by: 5) {
-                for l in stride(from: 0.0, through: 100, by: 5) {
-                    let color = NSColor(hue: h, saturation: s, lightness: l, alpha: 1.0)
+    func testUIColorRGBRangeTolerance() {
+        for h in stride(from: 0, through: 360, by: 5) {
+            for s in stride(from: 0, through: 100, by: 5) {
+                for l in stride(from: 0, through: 100, by: 5) {
+                    let color = UIColor(hue: Double(h), saturation: Double(s), lightness: Double(l), alpha: 1.0)
 
                     XCTAssertNotNil(color)
 
@@ -49,3 +51,5 @@ class AppKitTests: XCTestCase {
         }
     }
 }
+
+#endif
